+++ src/modules/utils.ts
@@ -48,6 +48,7 @@ export class FPythonScriptFiles {
     static readonly isDevmodeEnabled = "setup_code_completion/is_devmode_enabled";
     static readonly buildDocumentationToC = "documentation/build_toc";
     static readonly getDocPageContent = "documentation/get_page_content";
+    static readonly addSysPath = "add_sys_path";
 
     /** Get the absolute path to one of the scripts defined in this struct */
     static getAbsPath(file: string) {