+++ test/Test.py
@@ -24,8 +24,16 @@ def large_output():
         print(item)
     print('Done.')
 
+
 def message_box():
     result = unreal.EditorDialog().show_message("Title", "Hello World", unreal.AppMsgType.YES_NO_CANCEL, unreal.AppReturnType.YES)
     print(f"{result = }")
+
+
+def workspace_import():
+    import other_module
+    other_module.hello_world()
+
+
 
 log()